move dupe utils to util file
